requirements.txt
# Core
python-dotenv>=1.0.0
pandas>=2.0
psycopg2-binary>=2.9
alembic>=1.10
yfinance>=0.2.25
httpx>=0.24
beautifulsoup4>=4.12
python-dateutil>=2.8
uvicorn>=0.22
fastapi>=0.95

# Optional / baseline sentiment
#nltk>=3.8
#vaderSentiment>=3.3

# Dev & testing
pytest>=7.0
black>=24.3


# embeddings & vector DB
langchain>=0.0
chromadb>=0.3.28
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0 ; # optional: required for many HF models (CPU ok); adjust to your env
tqdm>=4.65

langfuse>=2.7.0
python-jose>=3.3
passlib[bcrypt]>=1.7

streamlit>=1.30
plotly>=5.18
apscheduler>=3.10
requests>=2.31

numpy>=1.21
pandas>=2.0
matplotlib>=3.7
scipy>=1.10

